make dns parser/writer public
