+++ Makefile.m32
@@ -63,6 +63,7 @@ install:
 	chmod u-w ${DESTDIR}${libdir}/$(LIB)
 	${INSTALL} -m 444 ${srcdir}/include/ares.h ${DESTDIR}${includedir}
 	${INSTALL} -m 444 ${srcdir}/include/ares_build.h ${DESTDIR}${includedir}
+	${INSTALL} -m 444 ${srcdir}/include/ares_dns_record.h ${DESTDIR}${includedir}
 	${INSTALL} -m 444 ${srcdir}/include/ares_rules.h ${DESTDIR}${includedir}
 	${INSTALL} -m 444 ${srcdir}/include/ares_version.h ${DESTDIR}${includedir}
 	(for man in $(MANPAGES); do \
