+++ src/lib/ares_private.h
@@ -122,7 +122,7 @@ typedef struct ares_rand_state ares_rand_state;
 #include "ares__htable_szvp.h"
 #include "ares__htable_asvp.h"
 #include "ares__buf.h"
-#include "ares_dns_record.h"
+#include "ares_dns_private.h"
 
 #ifndef HAVE_GETENV
 #  include "ares_getenv.h"
